RENAME config to webpackDllsConfig for clarity and easier search-ability
