+++ lib/ui/scripts/script.js
@@ -1,7 +1,7 @@
 import path from 'path';
 import webpack from 'webpack';
 
-import config from './config';
+import config from './webpackDllsConfig';
 
 const resolveLocal = dir => path.join(__dirname, dir);
 const webpackAsPromised = c =>
